Moved autonomus commands to a subpackage.
