+++ src/org/usfirst/frc/team3335/robot/commands/autonomous/AutoPlaceGear.java
@@ -1,4 +1,6 @@
-package org.usfirst.frc.team3335.robot.commands;
+package org.usfirst.frc.team3335.robot.commands.autonomous;
+
+import org.usfirst.frc.team3335.robot.commands.BallShiftLow;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
